Add solution for 2015, day 24
I had some struggles with generating all possible combinations. My
first attempt was to create a generator but once I started to use
recursion to attempt to solve the problem it became very clunky. Once
I moved away from using a generator things became a lot easier.

The first time solving part 1 it took several seconds to get the
solution. With a few modifications I got the time down to less than
200ms. I stored the best performing group, this was changed to just a
length and a quantum entanglement score. Then I pruned branches when
the number of packages exceeded the number of packages in the best
scoring group thus far.

The final optimisation was to stop finding combinations once the
packages left to make combinations with would no longer add up to the
target weight for a group.