Remove conversion to number for puzzle ID
When splitting the puzzle ID, each part was converted to a number.
This is entirely unnecessary as the rest of the code needs strings and
everywhere a `toString()` is needed to use part of the puzzle ID.

An additional benefit of no longer converting it to numbers is that
the `part` can be something non-numerical like `n-alt` to indicate
it is an alternative solution to part n of a puzzle.