Add solution for 2016, day 1
Part 1 was quickly done but part 2 gave me some trouble. Understanding
the challenge was the first hurdle, it took a while before I
understood what was meant by visited location. This is not just the
points you end up after following a direction but also any location
in between.

Once the challenge was clear a small modification to the solution for
part 1 was enough to get all the points from where we travelled into
a new direction. This made it possible to create line segments and to
find two intersecting line segments.