Refactor measurement code
The existing measurement code had a huge flaw. Module scoped variables
were not getting reset in between samples. For day 11 I had to rewrite
my solution to ensure it would measure properly but this is not
something I want to have to do. I want to be able to write the code
as I want it, without having to worry about introducing side-effects
to my measurements.

This new measurement code will spawn a new process for each sample.
Due to this the solution code will be loaded fresh ensuring there is
no data kept from the last run in module scoped variables. This will
solve the problem with the old code where data kept in module scoped
variable would solve challenges nearly instantaneous on
subsequent runs.

The `performance` file had a bit of an odd name so I renamed it to
something which fit better in the general scheme of things.